rename changes, minor bug fixes
