+++ src/constants.js
@@ -3,10 +3,10 @@ export const TICK_RATE = 3000;
 export const SCENES = ["day", "rain"];
 export const RAIN_CHANCE = 0.2;
 export const DAY_LENGTH = 60;
-export const NIGHT_LENGTH = 3;
+export const NIGHT_LENGTH = 5;
 export const getNextHugerTime = clock =>
-    Math.floor(Math.random() * 3) + 5 + clock;
+    Math.floor(Math.random() * 3) + 8 + clock;
 export const getNextDieTime = clock =>
-    Math.floor(Math.random() * 2) + 3 + clock;
+    Math.floor(Math.random() * 3) + 3 + clock;
 export const getNextPoopTime = clock =>
-    Math.floor(Math.random() * 3) + 4 + clock;
+    Math.floor(Math.random() * 3) + 8 + clock;
